# PERF: 20% faster pk attribute access